requirements.txt
matplotlib
numpy
biopython
tqdm
bitarray
pandas
seaborn